# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DevUtils Utility API",
    "description": "High-performance, developer-focused utility APIs for encoding/decoding, generation, parsing, and safe RE2 pattern testing.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current environment v1 base path"
    }
  ],
  "paths": {
    "/base64/encode": {
      "post": {
        "summary": "Encode plain text to Base64",
        "description": "Encodes standard UTF-8 string data into base64 format.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful conversion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Base64Result"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/base64/decode": {
      "post": {
        "summary": "Decode Base64 text to plain text",
        "description": "Decodes base64-encoded strings (standard or URL-safe) back into readable text.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful conversion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Base64Result"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/guid/generate": {
      "post": {
        "summary": "Generate batch UUIDs/GUIDs",
        "description": "Generates a batch of UUID v4 (random), UUID v7 (time-ordered), or custom prefixed identifiers.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GUIDRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/GUIDRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful batch generation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GUIDResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/qrcode/generate": {
      "post": {
        "summary": "Generate QR Code",
        "description": "Creates a PNG QR code representing the text or URL provided, returning base64-encoded visual formats.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful image creation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QRCodeResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/json/prettify": {
      "post": {
        "summary": "Prettify JSON data",
        "description": "Formats raw, compact JSON into beautified readable spacing.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JSONRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/JSONRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Formatted response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JSONFormattingResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/json/minify": {
      "post": {
        "summary": "Minify JSON data",
        "description": "Compacts JSON into a single line removing unnecessary whitespace.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Minified response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JSONFormattingResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/json/validate": {
      "post": {
        "summary": "Validate JSON format",
        "description": "Validates the correctness of a raw JSON schema structure.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JSONValidationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/yaml/json2yaml": {
      "post": {
        "summary": "Convert JSON to YAML",
        "description": "Converts strict or loose JSON payload into formatted YAML, preserving key order.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YAMLRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/YAMLRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "YAML output",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YAMLFormattingResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/yaml/yaml2json": {
      "post": {
        "summary": "Convert YAML to JSON",
        "description": "Converts YAML payload into formatted JSON, preserving key order.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YAMLRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/YAMLRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON output",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YAMLFormattingResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/yaml/prettify": {
      "post": {
        "summary": "Prettify YAML data",
        "description": "Formats raw YAML data into prettified spacing, with option to force quote styles.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YAMLRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/YAMLRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prettified YAML output",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YAMLFormattingResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/yaml/validate": {
      "post": {
        "summary": "Validate YAML format",
        "description": "Validates the syntax and correctness of raw YAML payloads.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YAMLRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/YAMLRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YAMLValidationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/sql/prettify": {
      "post": {
        "summary": "Prettify SQL",
        "description": "Formats raw SQL payload with correct indentation.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SQLRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/SQLRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Formatted SQL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SQLResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/sql/minify": {
      "post": {
        "summary": "Minify SQL",
        "description": "Compresses SQL query by removing unnecessary whitespace and newlines.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SQLRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/SQLRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Minified SQL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SQLResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/url/encode": {
      "post": {
        "summary": "URL Encode query text",
        "description": "Converts plain text into a percent-encoded query parameter format.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URL-safe response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/URLFormattingResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/url/decode": {
      "post": {
        "summary": "URL Decode percent-encoded text",
        "description": "Decodes percent-encoded character strings back to clear text.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Decoded response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/URLFormattingResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/url/parse": {
      "post": {
        "summary": "Dissect and Parse a URL",
        "description": "Extracts the Scheme, Host, Path, and individual Query Parameters of a full URL or query string.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parsed details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsedURLResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/regex/test": {
      "post": {
        "summary": "Test a regular expression",
        "description": "Compiles a Go RE2 expression, tests it against sample text, and returns highlighted segments, matches, and capture groups. Results are bounded to the first 100 matches.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegexTestRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/RegexTestRequest"
              },
              "encoding": {
                "flags": {
                  "style": "form",
                  "explode": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "RE2 match report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegexTestResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/jwt/decode": {
      "post": {
        "summary": "Decode JSON Web Token",
        "description": "Locally inspects standard JSON Web Token (JWT) Header, Payload, Claims, and Expiration parameters without checking signatures.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JWTRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/JWTRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parsed JWT details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecodedJWTResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/epoch/convert": {
      "post": {
        "summary": "Convert Epoch Timestamp / DateTime",
        "description": "Transforms numeric timestamps (seconds/millis) or human-readable dates into UTC, Local, Relative, and reciprocal epoch formats.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpochRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/EpochRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Converted representation metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpochResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/html-entity/encode": {
      "post": {
        "summary": "Encode HTML entities",
        "description": "Converts special characters to their HTML entity equivalents (& to &amp;, < to &lt;, etc.).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful encoding",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTMLEntityResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/html-entity/decode": {
      "post": {
        "summary": "Decode HTML entities",
        "description": "Converts HTML entities back to their original characters.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful decoding",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTMLEntityResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/diff/text": {
      "post": {
        "summary": "Compare two text blocks",
        "description": "Computes a line-by-line diff between two text blocks, returning added, removed, and unchanged lines with statistics.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiffRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/DiffRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful comparison",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    },
    "/cron/parse": {
      "post": {
        "summary": "Parse Cron Expression",
        "description": "Parses standard 5-field, 6-field (with seconds), or descriptor (e.g. @daily) cron expressions, returning a human-readable description and a list of upcoming execution times.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CronRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parsed cron schedule analysis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CronResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400Error"
          },
          "429": {
            "$ref": "#/components/responses/429Error"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "400Error": {
        "description": "Malformed payloads or missing variables",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenericError"
            }
          }
        }
      },
      "429Error": {
        "description": "Too many requests to endpoints (rate limited)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenericError"
            }
          }
        }
      }
    },
    "schemas": {
      "TextPayload": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string",
            "description": "Base utility string contents input payload"
          }
        }
      },
...
      "Base64Result": {
        "type": "object",
        "properties": {
          "result": {
            "type": "string",
            "description": "Standard base64 encoded or decoded text output"
          },
          "original": {
            "type": "string",
            "description": "Source text provided during execution request"
          },
          "operation": {
            "type": "string",
            "enum": [
              "Encode",
              "Decode"
            ]
          }
        }
      },
      "GUIDRequest": {
        "type": "object",
        "properties": {
          "count": {
            "type": "string",
            "description": "Batch quantity size (clamped 10 - 100). Default '10'.",
            "default": "10"
          },
          "type": {
            "type": "string",
            "description": "Type of ID generation algorithm schema style",
            "enum": [
              "v4",
              "v7",
              "prefixed",
              "prefixed_guid"
            ],
            "default": "v4"
          },
          "prefix": {
            "type": "string",
            "description": "Optional prefix text tag (used with prefixed/prefixed_guid modes)"
          }
        }
      },
      "GUIDResult": {
        "type": "object",
        "properties": {
          "guids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array list containing generated identifiers"
          },
          "count": {
            "type": "integer",
            "description": "Total amount generated"
          },
          "type": {
            "type": "string",
            "description": "ID format style"
          },
          "prefix": {
            "type": "string",
            "description": "Requested prefix value"
          }
        }
      },
      "QRCodeResult": {
        "type": "object",
        "properties": {
          "image_url": {
            "type": "string",
            "description": "Standard inline image Data URL (data:image/png;base64,...)"
          },
          "text": {
            "type": "string",
            "description": "Input payload encoded within the QR code graphic grid"
          },
          "raw_base64": {
            "type": "string",
            "description": "Base64 payload representation minus HTTP url decoration"
          }
        }
      },
...
    }
  }
}